requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
streamlit
plotly
networkx
pgmpy
numpy
pandas
tensorflow
xgboost
scikit-learn
pulp
youtube-transcript-api
yt-dlp
folium
requests
pytest
rpy2
geopy
faker
feedparser
